+++ src/Paynow/Exception/PaynowException.php
@@ -6,9 +6,10 @@ use Exception;
 
 class PaynowException extends Exception
 {
+    /** @var Error[]|null */
     private $errors;
 
-    public function __construct($message = '', $code = 0, $body = null)
+    public function __construct(string $message, int $code = 0, ?string $body = null)
     {
         parent::__construct($message, $code, null);
 
@@ -18,15 +19,14 @@ class PaynowException extends Exception
                 foreach ($json->errors as $error) {
                     $this->errors[] = new Error($error->errorType, $error->message);
                 }
-                $this->errors = $json->errors;
             }
         }
     }
 
     /**
-     * @return mixed
+     * @return Error[]|null
      */
-    public function getErrors()
+    public function getErrors(): array
     {
         return $this->errors;
     }