+++ tests/NotificationTest.php
@@ -2,6 +2,7 @@
 
 namespace Paynow\Tests;
 
+use InvalidArgumentException;
 use Paynow\Exception\SignatureVerificationException;
 use Paynow\Notification;
 
@@ -9,6 +10,9 @@ class NotificationTest extends TestCase
 {
     /**
      * @dataProvider requestsToTest
+     * @param $payload
+     * @param $headers
+     * @throws SignatureVerificationException
      */
     public function testVerifyPayloadSuccessfully($payload, $headers)
     {
@@ -27,11 +31,11 @@ class NotificationTest extends TestCase
         return [
             [
                 $payload,
-                ['Signature' => 'UZgTT6iSv174R/OyQ2DWRCE9UCmvdXDS8rbQQcjk+AA=']
+                ['Signature' => 'Aq/VmN15rtjVbuy9F7Yw+Ym76H+VZjVSuHGpg4dwitY=']
             ],
             [
                 $payload,
-                ['signature' => 'UZgTT6iSv174R/OyQ2DWRCE9UCmvdXDS8rbQQcjk+AA=']
+                ['signature' => 'Aq/VmN15rtjVbuy9F7Yw+Ym76H+VZjVSuHGpg4dwitY=']
             ]
         ];
     }
@@ -41,7 +45,7 @@ class NotificationTest extends TestCase
         // given
         $this->expectException(SignatureVerificationException::class);
         $payload = $this->loadData('notification.json', true);
-        $headers = ['Signature' => 'Aq/VmN15rtjVbuy9F7Yw+Ym76H+VZjVSuHGpg4dwitY='];
+        $headers = ['Signature' => 'Wq/V2N15rtjVbuy9F7Yw+Ym76H+VZjVSuHGpg4dwitY='];
 
         // when
         new Notification('s3ecret-k3y', $payload, $headers);
@@ -52,7 +56,7 @@ class NotificationTest extends TestCase
     public function testShouldThrowExceptionOnMissingPayload()
     {
         // given
-        $this->expectException(\InvalidArgumentException::class);
+        $this->expectException(InvalidArgumentException::class);
         $payload = null;
         $headers = [];
 
@@ -65,7 +69,7 @@ class NotificationTest extends TestCase
     public function testShouldThrowExceptionOnMissingPayloadHeaders()
     {
         // given
-        $this->expectException(\InvalidArgumentException::class);
+        $this->expectException(InvalidArgumentException::class);
         $payload = $this->loadData('notification.json', true);
         $headers = null;