+++ tests/Util/SignatureCalculatorTest.php
@@ -2,6 +2,7 @@
 
 namespace Paynow\Tests\Util;
 
+use InvalidArgumentException;
 use Paynow\Tests\TestCase;
 use Paynow\Util\SignatureCalculator;
 
@@ -10,7 +11,7 @@ class SignatureCalculatorTest extends TestCase
     public function testNotValidSuccessfully()
     {
         // given + when
-        $signatureCalculator = new SignatureCalculator('InvalidSecretKey', ['key' => 'value']);
+        $signatureCalculator = new SignatureCalculator('InvalidSecretKey', json_encode(['key' => 'value']));
 
         // then
         $this->assertNotEquals('hash', $signatureCalculator->getHash());
@@ -21,23 +22,23 @@ class SignatureCalculatorTest extends TestCase
         // given + when
         $signatureCalculator = new SignatureCalculator(
             'a621a1fb-b4d8-48ba-a6a3-2a28ed61f605',
-            [
+            json_encode([
                 'key1' => 'value1',
-                'key2' => 'val/ue2',
-            ]
+                'key2' => 'value2',
+            ])
         );
 
         // then
-        $this->assertEquals('vnZs3ZaiBGFd9R6qEfgKO1psx+S2P9sFIcY4AO5hJ78=', $signatureCalculator->getHash());
+        $this->assertEquals('rFAkhfbUFRn4bTR82qb742Mwy34g/CSi8frEHciZhCU=', $signatureCalculator->getHash());
     }
 
     public function testExceptionForEmptyData()
     {
         // given
-        $this->expectException(\InvalidArgumentException::class);
+        $this->expectException(InvalidArgumentException::class);
 
         // when
-        $signatureCalculator = new SignatureCalculator('a621a1fb-b4d8-48ba-a6a3-2a28ed61f605', []);
+        $signatureCalculator = new SignatureCalculator('a621a1fb-b4d8-48ba-a6a3-2a28ed61f605', "");
 
         // then
     }